manipmongo: Reduce migration review noise
Purpose: Make the mongo-driver migration easier to review by removing non-essential indirection and restoring legacy-facing details that do not need to change in this PR.

Includes:
- restore existing trace names for manipmongo CRUD and count operations
- simplify forced read filter setup and drop an unnecessary local variable
- replace trivial defaultContext/queryContext wrapper methods with direct helper usage
- move sharding and previous-retriever helpers out of the middle of manipulator.go into utils.go
- update helper and sharder tests to match the cleanup without changing migration behavior
