+++ manipmongo/utils.go
@@ -16,7 +16,9 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"maps"
 	"net"
+	neturl "net/url"
 	"strings"
 	"time"
 
@@ -24,8 +26,12 @@ import (
 	"github.com/opentracing/opentracing-go/log"
 	"go.acuvity.ai/elemental"
 	"go.acuvity.ai/manipulate"
+	"go.acuvity.ai/manipulate/internal/backoff"
 	bson "go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
+	"go.mongodb.org/mongo-driver/v2/mongo/options"
+	"go.mongodb.org/mongo-driver/v2/mongo/readpref"
+	"go.mongodb.org/mongo-driver/v2/mongo/writeconcern"
 )
 
 const (
@@ -34,12 +40,119 @@ const (
 	errInvalidQueryBadRegex     = "$regex has to be a string"
 )
 
-func applyOrdering(order []string, spec elemental.AttributeSpecifiable) []string {
+type countRes struct {
+	Count int `bson:"_count"`
+}
 
+type mongoOperationContextError struct {
+	Err error
+}
+
+type mongoQueryError struct {
+	Code    int
+	Message string
+	Err     error
+}
+
+func (e *mongoOperationContextError) Error() string {
+	if e == nil || e.Err == nil {
+		return ""
+	}
+	return e.Err.Error()
+}
+
+func (e *mongoOperationContextError) Unwrap() error {
+	if e == nil {
+		return nil
+	}
+	return e.Err
+}
+
+func (e *mongoQueryError) Error() string {
+	if e == nil {
+		return ""
+	}
+	if e.Err != nil {
+		return e.Err.Error()
+	}
+	return e.Message
+}
+
+func (e *mongoQueryError) Unwrap() error {
+	if e == nil {
+		return nil
+	}
+	return e.Err
+}
+
+func HandleQueryErrorMongo(err error) error {
+	return HandleQueryError(err)
+}
+
+// HandleQueryError handles the provided upstream error returned by Mongo by returning a corresponding manipulate error type.
+func HandleQueryError(err error) error {
+	if err == nil {
+		return nil
+	}
+
+	var operationCtxErr *mongoOperationContextError
+	if errors.As(err, &operationCtxErr) {
+		return manipulate.ErrCannotExecuteQuery{Err: operationCtxErr.Err}
+	}
+
+	if errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled) {
+		return manipulate.ErrCannotCommunicate{Err: err}
+	}
+
+	var netErr net.Error
+	if errors.As(err, &netErr) {
+		return manipulate.ErrCannotCommunicate{Err: err}
+	}
+
+	if errors.Is(err, mongo.ErrNoDocuments) {
+		return manipulate.ErrObjectNotFound{Err: fmt.Errorf("cannot find the object for the given ID")}
+	}
+
+	if mongo.IsDuplicateKeyError(err) {
+		return manipulate.ErrConstraintViolation{Err: fmt.Errorf("duplicate key")}
+	}
+
+	if mongo.IsTimeout(err) || mongo.IsNetworkError(err) || isConnectionError(err) {
+		return manipulate.ErrCannotCommunicate{Err: err}
+	}
+
+	if ok, invalidErr := invalidQuery(err); ok {
+		return invalidErr
+	}
+
+	// see https://github.com/mongodb/mongo/blob/master/src/mongo/base/error_codes.err
+	switch getErrorCode(err) {
+	case 6, 7, 71, 74, 91, 109, 189, 202, 216, 262, 10107, 13436, 13435, 11600, 11602:
+		// HostUnreachable
+		// HostNotFound
+		// ReplicaSetNotFound
+		// NodeNotFound
+		// ConfigurationInProgress
+		// ShutdownInProgress
+		// PrimarySteppedDown
+		// NetworkInterfaceExceededTimeLimit
+		// ElectionInProgress
+		// ExceededTimeLimit
+		// NotMaster
+		// NotMasterOrSecondary
+		// NotMasterNoSlaveOk
+		// InterruptedAtShutdown
+		// InterruptedDueToStepDown
+		return manipulate.ErrCannotCommunicate{Err: err}
+	default:
+		return manipulate.ErrCannotExecuteQuery{Err: err}
+	}
+}
+
+func applyOrdering(order []string, spec elemental.AttributeSpecifiable) []string {
 	o := []string{} // nolint: prealloc
 
 	for _, f := range order {
-
 		if f == "" {
 			continue
 		}
@@ -68,6 +181,338 @@ func applyOrdering(order []string, spec elemental.AttributeSpecifiable) []string
 	}
 
 	return o
+}
+
+func redactMongoURI(raw string) string {
+	parsed, err := neturl.Parse(raw)
+	if err != nil || parsed == nil || parsed.User == nil {
+		return raw
+	}
+
+	username := parsed.User.Username()
+	if username == "" {
+		username = "redacted"
+	}
+
+	if _, hasPassword := parsed.User.Password(); hasPassword {
+		parsed.User = neturl.UserPassword(username, "REDACTED")
+		return parsed.String()
+	}
+
+	parsed.User = neturl.User(username)
+	return parsed.String()
+}
+
+func countFromResults(res []*countRes) (int, error) {
+	switch len(res) {
+	case 0:
+		return 0, nil
+	case 1:
+		return res[0].Count, nil
+	default:
+		return 0, manipulate.ErrCannotExecuteQuery{Err: fmt.Errorf("count: invalid count result len: %d. must be 1", len(res))}
+	}
+}
+
+func contextWithDefaultTimeout(ctx context.Context, timeout time.Duration) (context.Context, context.CancelFunc) {
+	return context.WithTimeout(ctx, effectiveTimeout(timeout))
+}
+
+func effectiveTimeout(timeout time.Duration) time.Duration {
+	if timeout > 0 {
+		return timeout
+	}
+	return defaultOperationTimeout
+}
+
+func maxTimeFromContext(ctx context.Context, defaultTimeout time.Duration) (time.Duration, error) {
+	if err := ctx.Err(); err != nil {
+		return 0, manipulate.ErrCannotBuildQuery{Err: err}
+	}
+
+	d, ok := ctx.Deadline()
+	if !ok {
+		return effectiveTimeout(defaultTimeout), nil
+	}
+	return time.Until(d), nil
+}
+
+func mongoOperationContext(ctx context.Context, defaultTimeout time.Duration) (context.Context, context.CancelFunc, error) {
+	maxTime, err := maxTimeFromContext(ctx, defaultTimeout)
+	if err != nil {
+		return nil, nil, err
+	}
+	if maxTime <= 0 {
+		maxTime = time.Nanosecond
+	}
+	queryCtx, cancel := context.WithTimeout(ctx, maxTime)
+	return queryCtx, cancel, nil
+}
+
+func wrapMongoOperationContextError(parentCtx context.Context, queryCtx context.Context, err error) error {
+	if err == nil {
+		return nil
+	}
+	if (errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled)) && queryCtx != nil && queryCtx.Err() != nil && parentCtx != nil && parentCtx.Err() == nil {
+		return &mongoOperationContextError{Err: queryCtx.Err()}
+	}
+	return err
+}
+
+func identifierFilterElement(identifier string) bson.E {
+	if oid, err := bson.ObjectIDFromHex(identifier); err == nil {
+		return bson.E{Key: "_id", Value: oid}
+	}
+	return bson.E{Key: "_id", Value: identifier}
+}
+
+func composeAndFilter(filter bson.D, andFilters ...bson.D) bson.D {
+	if len(andFilters) == 0 {
+		return filter
+	}
+
+	clauses := make([]bson.D, 0, len(andFilters)+1)
+	for _, clause := range andFilters {
+		if len(clause) == 0 {
+			continue
+		}
+		clauses = append(clauses, clause)
+	}
+	clauses = append(clauses, filter)
+
+	return bson.D{{Key: "$and", Value: clauses}}
+}
+
+func convertUpsertOperationsToMongo(operations any) (bson.M, error) {
+	switch typed := operations.(type) {
+	case bson.M:
+		if typed == nil {
+			return bson.M{}, nil
+		}
+		return typed, nil
+	default:
+		return nil, fmt.Errorf("upsert operations must be bson.M")
+	}
+}
+
+func mergeSetOnInsertOperationsMongo(dst bson.M, soi any) error {
+	converted, err := convertSetOnInsertToMongoMap(soi)
+	if err != nil {
+		return err
+	}
+	maps.Copy(dst, converted)
+	return nil
+}
+
+func convertSetOnInsertToMongoMap(soi any) (bson.M, error) {
+	switch typed := soi.(type) {
+	case bson.M:
+		if typed == nil {
+			return bson.M{}, nil
+		}
+		return typed, nil
+	default:
+		return nil, fmt.Errorf("$setOnInsert operations must be bson.M")
+	}
+}
+
+func convertReadPreferenceMongo(c manipulate.ReadConsistency) *readpref.ReadPref {
+	switch c {
+	case manipulate.ReadConsistencyEventual:
+		// mgo.Eventual is effectively Nearest, but without session stickiness
+		// across sequential reads. The official driver does not expose that
+		// stickiness behavior as a read preference, so Nearest is the closest
+		// semantic match.
+		return readpref.Nearest()
+	case manipulate.ReadConsistencyMonotonic:
+		// Monotonic has no exact equivalent in mongo-driver read preferences.
+		// PrimaryPreferred is the safest approximation for read-after-write behavior.
+		return readpref.PrimaryPreferred()
+	case manipulate.ReadConsistencyNearest:
+		return readpref.Nearest()
+	case manipulate.ReadConsistencyStrong:
+		return readpref.Primary()
+	case manipulate.ReadConsistencyWeakest:
+		return readpref.SecondaryPreferred()
+	default:
+		return nil
+	}
+}
+
+func convertWriteConcernMongo(c manipulate.WriteConsistency) *writeconcern.WriteConcern {
+	switch c {
+	case manipulate.WriteConsistencyNone:
+		return writeconcern.Unacknowledged()
+	case manipulate.WriteConsistencyStrong:
+		return writeconcern.Majority()
+	case manipulate.WriteConsistencyStrongest:
+		journaled := writeconcern.Journaled()
+		journaled.W = writeconcern.WCMajority
+		return journaled
+	default:
+		return writeconcern.W1()
+	}
+}
+
+func runQueryMongo(mctx manipulate.Context, operationFunc func() (any, error), baseRetryInfo RetryInfo) (any, error) {
+	var try int
+	for {
+		out, err := operationFunc()
+		if err == nil {
+			return out, nil
+		}
+
+		err = HandleQueryErrorMongo(err)
+		if !manipulate.IsCannotCommunicateError(err) {
+			return out, err
+		}
+
+		baseRetryInfo.try = try
+		baseRetryInfo.err = err
+		baseRetryInfo.mctx = mctx
+
+		if rf := mctx.RetryFunc(); rf != nil {
+			if rerr := rf(baseRetryInfo); rerr != nil {
+				return nil, rerr
+			}
+		} else if baseRetryInfo.defaultRetryFunc != nil {
+			if rerr := baseRetryInfo.defaultRetryFunc(baseRetryInfo); rerr != nil {
+				return nil, rerr
+			}
+		}
+
+		select {
+		case <-mctx.Context().Done():
+			return nil, manipulate.ErrCannotExecuteQuery{Err: mctx.Context().Err()}
+		default:
+		}
+
+		deadline, _ := mctx.Context().Deadline()
+		time.Sleep(backoff.NextWithCurve(try, deadline, defaultBackoffCurve))
+		try++
+	}
+}
+
+func newMongo(url string, db string, opts ...Option) (manipulate.TransactionalManipulator, error) {
+	cfg := newConfig()
+	for _, o := range opts {
+		o(cfg)
+	}
+
+	if cfg.poolLimit < 0 {
+		panic(fmt.Errorf("manipmongo: invalid connection pool limit %d: must be greater than or equal to 0", cfg.poolLimit))
+	}
+
+	validatedForcedReadFilter := bson.D{}
+	if len(cfg.forcedReadFilter) > 0 {
+		validatedForcedReadFilter = append(validatedForcedReadFilter, cfg.forcedReadFilter...)
+		if _, err := bson.Marshal(validatedForcedReadFilter); err != nil {
+			panic(fmt.Errorf("manipmongo: invalid forced read filter: %w", err))
+		}
+	}
+
+	mongoPoolLimit := cfg.poolLimit
+	if mongoPoolLimit == 0 {
+		mongoPoolLimit = newConfig().poolLimit
+	}
+
+	clientOpts := options.Client().
+		ApplyURI(url).
+		SetMaxPoolSize(uint64(mongoPoolLimit)).
+		SetConnectTimeout(cfg.connectTimeout)
+
+	if cfg.clientTimeout > 0 {
+		clientOpts.SetTimeout(cfg.clientTimeout)
+	}
+
+	if cfg.username != "" || cfg.password != "" || cfg.authsource != "" {
+		clientOpts.SetAuth(options.Credential{
+			Username:   cfg.username,
+			Password:   cfg.password,
+			AuthSource: cfg.authsource,
+		})
+	}
+
+	if cfg.tlsConfig != nil {
+		clientOpts.SetTLSConfig(cfg.tlsConfig)
+	}
+
+	if rp := convertReadPreferenceMongo(cfg.readConsistency); rp != nil {
+		clientOpts.SetReadPreference(rp)
+	}
+	if wc := convertWriteConcernMongo(cfg.writeConsistency); wc != nil {
+		clientOpts.SetWriteConcern(wc)
+	}
+
+	client, err := mongoConnectFn(clientOpts)
+	if err != nil {
+		return nil, fmt.Errorf("cannot connect to mongo url '%s': %w", redactMongoURI(url), err)
+	}
+
+	ctx, cancel := contextWithDefaultTimeout(context.Background(), cfg.connectTimeout)
+	defer cancel()
+	if err := mongoPingFn(client, ctx); err != nil {
+		disconnectCtx, disconnectCancel := contextWithDefaultTimeout(context.Background(), cfg.connectTimeout)
+		defer disconnectCancel()
+		_ = mongoDisconnectFn(client, disconnectCtx)
+		return nil, fmt.Errorf("cannot ping mongo url '%s': %w", redactMongoURI(url), err)
+	}
+
+	return &mongoManipulator{
+		client:                  client,
+		dbName:                  db,
+		sharder:                 cfg.sharder,
+		defaultRetryFunc:        cfg.defaultRetryFunc,
+		forcedReadFilter:        validatedForcedReadFilter,
+		attributeEncrypter:      cfg.attributeEncrypter,
+		explain:                 cfg.explain,
+		attributeSpecifiers:     cfg.attributeSpecifiers,
+		operationTimeout:        cfg.operationTimeout,
+		defaultReadConsistency:  cfg.readConsistency,
+		defaultWriteConsistency: cfg.writeConsistency,
+	}, nil
+}
+
+func makePreviousRetriever(ctx context.Context, coll *mongo.Collection, defaultTimeout time.Duration) func(id bson.ObjectID) (bson.M, error) {
+	return func(id bson.ObjectID) (bson.M, error) {
+		queryCtx, cancel, err := mongoOperationContext(ctx, defaultTimeout)
+		if err != nil {
+			return nil, err
+		}
+		defer cancel()
+
+		doc := bson.M{}
+		if err := coll.FindOne(queryCtx, bson.D{{Key: "_id", Value: id}}).Decode(&doc); err != nil {
+			return nil, wrapMongoOperationContextError(ctx, queryCtx, err)
+		}
+		return doc, nil
+	}
+}
+
+func makeShardingManyFilter(m *mongoManipulator, mctx manipulate.Context, identity elemental.Identity) (bson.D, error) {
+	if m.sharder == nil {
+		return nil, nil
+	}
+
+	filter, err := m.sharder.FilterMany(m, mctx, identity)
+	if err != nil {
+		return nil, manipulate.ErrCannotBuildQuery{Err: fmt.Errorf("cannot compute sharding filter: %w", err)}
+	}
+
+	return filter, nil
+}
+
+func makeShardingOneFilter(m *mongoManipulator, mctx manipulate.Context, object elemental.Identifiable) (bson.D, error) {
+	if m.sharder == nil {
+		return nil, nil
+	}
+
+	filter, err := m.sharder.FilterOne(m, mctx, object)
+	if err != nil {
+		return nil, manipulate.ErrCannotBuildQuery{Err: fmt.Errorf("cannot compute sharding filter: %w", err)}
+	}
+
+	return filter, nil
 }
 
 func makeNamespaceFilter(mctx manipulate.Context) bson.D {
@@ -228,90 +673,6 @@ func spanErr(sp opentracing.Span, err error) error {
 	}
 
 	return err
-}
-
-// HandleQueryError handles the provided upstream error returned by Mongo by returning a corresponding manipulate error type.
-func HandleQueryError(err error) error {
-
-	if err == nil {
-		return nil
-	}
-
-	var operationCtxErr *mongoOperationContextError
-	if errors.As(err, &operationCtxErr) {
-		return manipulate.ErrCannotExecuteQuery{Err: operationCtxErr.Err}
-	}
-
-	if errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled) {
-		return manipulate.ErrCannotCommunicate{Err: err}
-	}
-
-	var netErr net.Error
-	if errors.As(err, &netErr) {
-		return manipulate.ErrCannotCommunicate{Err: err}
-	}
-
-	if errors.Is(err, mongo.ErrNoDocuments) {
-		return manipulate.ErrObjectNotFound{Err: fmt.Errorf("cannot find the object for the given ID")}
-	}
-
-	if mongo.IsDuplicateKeyError(err) {
-		return manipulate.ErrConstraintViolation{Err: fmt.Errorf("duplicate key")}
-	}
-
-	if mongo.IsTimeout(err) || mongo.IsNetworkError(err) || isConnectionError(err) {
-		return manipulate.ErrCannotCommunicate{Err: err}
-	}
-
-	if ok, invalidErr := invalidQuery(err); ok {
-		return invalidErr
-	}
-
-	// see https://github.com/mongodb/mongo/blob/master/src/mongo/base/error_codes.err
-	switch getErrorCode(err) {
-	case 6, 7, 71, 74, 91, 109, 189, 202, 216, 262, 10107, 13436, 13435, 11600, 11602:
-		// HostUnreachable
-		// HostNotFound
-		// ReplicaSetNotFound
-		// NodeNotFound
-		// ConfigurationInProgress
-		// ShutdownInProgress
-		// PrimarySteppedDown
-		// NetworkInterfaceExceededTimeLimit
-		// ElectionInProgress
-		// ExceededTimeLimit
-		// NotMaster
-		// NotMasterOrSecondary
-		// NotMasterNoSlaveOk
-		// InterruptedAtShutdown
-		// InterruptedDueToStepDown
-		return manipulate.ErrCannotCommunicate{Err: err}
-	default:
-		return manipulate.ErrCannotExecuteQuery{Err: err}
-	}
-}
-
-type mongoQueryError struct {
-	Code    int
-	Message string
-	Err     error
-}
-
-func (e *mongoQueryError) Error() string {
-	if e == nil {
-		return ""
-	}
-	if e.Err != nil {
-		return e.Err.Error()
-	}
-	return e.Message
-}
-
-func (e *mongoQueryError) Unwrap() error {
-	if e == nil {
-		return nil
-	}
-	return e.Err
 }
 
 func getErrorCode(err error) int {
